Move donation policy to players
